+++ fc-open-20210406/alibabacloud_fc_open20210406/models.py
@@ -410,6 +410,63 @@ class CustomDNS(TeaModel):
         return self
 
 
+class CustomHealthCheckConfig(TeaModel):
+    def __init__(
+        self,
+        failure_threshold: int = None,
+        http_get_url: str = None,
+        initial_delay_seconds: int = None,
+        period_seconds: int = None,
+        success_threshold: int = None,
+        timeout_seconds: int = None,
+    ):
+        self.failure_threshold = failure_threshold
+        self.http_get_url = http_get_url
+        self.initial_delay_seconds = initial_delay_seconds
+        self.period_seconds = period_seconds
+        self.success_threshold = success_threshold
+        self.timeout_seconds = timeout_seconds
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.failure_threshold is not None:
+            result['failureThreshold'] = self.failure_threshold
+        if self.http_get_url is not None:
+            result['httpGetUrl'] = self.http_get_url
+        if self.initial_delay_seconds is not None:
+            result['initialDelaySeconds'] = self.initial_delay_seconds
+        if self.period_seconds is not None:
+            result['periodSeconds'] = self.period_seconds
+        if self.success_threshold is not None:
+            result['successThreshold'] = self.success_threshold
+        if self.timeout_seconds is not None:
+            result['timeoutSeconds'] = self.timeout_seconds
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('failureThreshold') is not None:
+            self.failure_threshold = m.get('failureThreshold')
+        if m.get('httpGetUrl') is not None:
+            self.http_get_url = m.get('httpGetUrl')
+        if m.get('initialDelaySeconds') is not None:
+            self.initial_delay_seconds = m.get('initialDelaySeconds')
+        if m.get('periodSeconds') is not None:
+            self.period_seconds = m.get('periodSeconds')
+        if m.get('successThreshold') is not None:
+            self.success_threshold = m.get('successThreshold')
+        if m.get('timeoutSeconds') is not None:
+            self.timeout_seconds = m.get('timeoutSeconds')
+        return self
+
+
 class CustomRuntimeConfig(TeaModel):
     def __init__(
         self,
@@ -1472,6 +1529,86 @@ class NASConfig(TeaModel):
         return self
 
 
+class OSSMountConfigMountPoints(TeaModel):
+    def __init__(
+        self,
+        bucket_name: str = None,
+        endpoint: str = None,
+        mount_dir: str = None,
+        read_only: bool = None,
+    ):
+        self.bucket_name = bucket_name
+        self.endpoint = endpoint
+        self.mount_dir = mount_dir
+        self.read_only = read_only
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.bucket_name is not None:
+            result['bucketName'] = self.bucket_name
+        if self.endpoint is not None:
+            result['endpoint'] = self.endpoint
+        if self.mount_dir is not None:
+            result['mountDir'] = self.mount_dir
+        if self.read_only is not None:
+            result['readOnly'] = self.read_only
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('bucketName') is not None:
+            self.bucket_name = m.get('bucketName')
+        if m.get('endpoint') is not None:
+            self.endpoint = m.get('endpoint')
+        if m.get('mountDir') is not None:
+            self.mount_dir = m.get('mountDir')
+        if m.get('readOnly') is not None:
+            self.read_only = m.get('readOnly')
+        return self
+
+
+class OSSMountConfig(TeaModel):
+    def __init__(
+        self,
+        mount_points: List[OSSMountConfigMountPoints] = None,
+    ):
+        self.mount_points = mount_points
+
+    def validate(self):
+        if self.mount_points:
+            for k in self.mount_points:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['mountPoints'] = []
+        if self.mount_points is not None:
+            for k in self.mount_points:
+                result['mountPoints'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.mount_points = []
+        if m.get('mountPoints') is not None:
+            for k in m.get('mountPoints'):
+                temp_model = OSSMountConfigMountPoints()
+                self.mount_points.append(temp_model.from_map(k))
+        return self
+
+
 class OSSTriggerKey(TeaModel):
     def __init__(
         self,
@@ -2459,6 +2596,209 @@ class VendorConfig(TeaModel):
         return self
 
 
+class ClaimGPUInstanceHeaders(TeaModel):
+    def __init__(
+        self,
+        common_headers: Dict[str, str] = None,
+        x_fc_account_id: str = None,
+        x_fc_date: str = None,
+        x_fc_trace_id: str = None,
+    ):
+        self.common_headers = common_headers
+        self.x_fc_account_id = x_fc_account_id
+        self.x_fc_date = x_fc_date
+        self.x_fc_trace_id = x_fc_trace_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.common_headers is not None:
+            result['commonHeaders'] = self.common_headers
+        if self.x_fc_account_id is not None:
+            result['X-Fc-Account-Id'] = self.x_fc_account_id
+        if self.x_fc_date is not None:
+            result['X-Fc-Date'] = self.x_fc_date
+        if self.x_fc_trace_id is not None:
+            result['X-Fc-Trace-Id'] = self.x_fc_trace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('commonHeaders') is not None:
+            self.common_headers = m.get('commonHeaders')
+        if m.get('X-Fc-Account-Id') is not None:
+            self.x_fc_account_id = m.get('X-Fc-Account-Id')
+        if m.get('X-Fc-Date') is not None:
+            self.x_fc_date = m.get('X-Fc-Date')
+        if m.get('X-Fc-Trace-Id') is not None:
+            self.x_fc_trace_id = m.get('X-Fc-Trace-Id')
+        return self
+
+
+class ClaimGPUInstanceRequest(TeaModel):
+    def __init__(
+        self,
+        disk_performance_level: str = None,
+        disk_size_gigabytes: str = None,
+        image_id: str = None,
+        instance_type: str = None,
+        internet_bandwidth_out: str = None,
+        password: str = None,
+        source_cidr_ip: str = None,
+        tcp_port_range: List[str] = None,
+        udp_port_range: List[str] = None,
+    ):
+        self.disk_performance_level = disk_performance_level
+        self.disk_size_gigabytes = disk_size_gigabytes
+        self.image_id = image_id
+        self.instance_type = instance_type
+        self.internet_bandwidth_out = internet_bandwidth_out
+        self.password = password
+        self.source_cidr_ip = source_cidr_ip
+        self.tcp_port_range = tcp_port_range
+        self.udp_port_range = udp_port_range
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.disk_performance_level is not None:
+            result['diskPerformanceLevel'] = self.disk_performance_level
+        if self.disk_size_gigabytes is not None:
+            result['diskSizeGigabytes'] = self.disk_size_gigabytes
+        if self.image_id is not None:
+            result['imageId'] = self.image_id
+        if self.instance_type is not None:
+            result['instanceType'] = self.instance_type
+        if self.internet_bandwidth_out is not None:
+            result['internetBandwidthOut'] = self.internet_bandwidth_out
+        if self.password is not None:
+            result['password'] = self.password
+        if self.source_cidr_ip is not None:
+            result['sourceCidrIp'] = self.source_cidr_ip
+        if self.tcp_port_range is not None:
+            result['tcpPortRange'] = self.tcp_port_range
+        if self.udp_port_range is not None:
+            result['udpPortRange'] = self.udp_port_range
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('diskPerformanceLevel') is not None:
+            self.disk_performance_level = m.get('diskPerformanceLevel')
+        if m.get('diskSizeGigabytes') is not None:
+            self.disk_size_gigabytes = m.get('diskSizeGigabytes')
+        if m.get('imageId') is not None:
+            self.image_id = m.get('imageId')
+        if m.get('instanceType') is not None:
+            self.instance_type = m.get('instanceType')
+        if m.get('internetBandwidthOut') is not None:
+            self.internet_bandwidth_out = m.get('internetBandwidthOut')
+        if m.get('password') is not None:
+            self.password = m.get('password')
+        if m.get('sourceCidrIp') is not None:
+            self.source_cidr_ip = m.get('sourceCidrIp')
+        if m.get('tcpPortRange') is not None:
+            self.tcp_port_range = m.get('tcpPortRange')
+        if m.get('udpPortRange') is not None:
+            self.udp_port_range = m.get('udpPortRange')
+        return self
+
+
+class ClaimGPUInstanceResponseBody(TeaModel):
+    def __init__(
+        self,
+        created_time: str = None,
+        instance_id: str = None,
+        public_ip: str = None,
+    ):
+        self.created_time = created_time
+        self.instance_id = instance_id
+        self.public_ip = public_ip
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.created_time is not None:
+            result['createdTime'] = self.created_time
+        if self.instance_id is not None:
+            result['instanceId'] = self.instance_id
+        if self.public_ip is not None:
+            result['publicIp'] = self.public_ip
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('createdTime') is not None:
+            self.created_time = m.get('createdTime')
+        if m.get('instanceId') is not None:
+            self.instance_id = m.get('instanceId')
+        if m.get('publicIp') is not None:
+            self.public_ip = m.get('publicIp')
+        return self
+
+
+class ClaimGPUInstanceResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: ClaimGPUInstanceResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        self.validate_required(self.headers, 'headers')
+        self.validate_required(self.status_code, 'status_code')
+        self.validate_required(self.body, 'body')
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = ClaimGPUInstanceResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class CreateAliasHeaders(TeaModel):
     def __init__(
         self,
@@ -2953,6 +3293,7 @@ class CreateFunctionRequest(TeaModel):
         code: Code = None,
         custom_container_config: CustomContainerConfig = None,
         custom_dns: CustomDNS = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         custom_runtime_config: CustomRuntimeConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
@@ -2973,6 +3314,7 @@ class CreateFunctionRequest(TeaModel):
         self.code = code
         self.custom_container_config = custom_container_config
         self.custom_dns = custom_dns
+        self.custom_health_check_config = custom_health_check_config
         self.custom_runtime_config = custom_runtime_config
         self.description = description
         self.environment_variables = environment_variables
@@ -2996,6 +3338,8 @@ class CreateFunctionRequest(TeaModel):
             self.custom_container_config.validate()
         if self.custom_dns:
             self.custom_dns.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.custom_runtime_config:
             self.custom_runtime_config.validate()
         if self.instance_lifecycle_config:
@@ -3015,6 +3359,8 @@ class CreateFunctionRequest(TeaModel):
             result['customContainerConfig'] = self.custom_container_config.to_map()
         if self.custom_dns is not None:
             result['customDNS'] = self.custom_dns.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.custom_runtime_config is not None:
             result['customRuntimeConfig'] = self.custom_runtime_config.to_map()
         if self.description is not None:
@@ -3060,6 +3406,9 @@ class CreateFunctionRequest(TeaModel):
         if m.get('customDNS') is not None:
             temp_model = CustomDNS()
             self.custom_dns = temp_model.from_map(m['customDNS'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('customRuntimeConfig') is not None:
             temp_model = CustomRuntimeConfig()
             self.custom_runtime_config = temp_model.from_map(m['customRuntimeConfig'])
@@ -3104,6 +3453,7 @@ class CreateFunctionResponseBody(TeaModel):
         created_time: str = None,
         custom_container_config: CustomContainerConfig = None,
         custom_dns: CustomDNS = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         custom_runtime_config: CustomRuntimeConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
@@ -3128,6 +3478,7 @@ class CreateFunctionResponseBody(TeaModel):
         self.created_time = created_time
         self.custom_container_config = custom_container_config
         self.custom_dns = custom_dns
+        self.custom_health_check_config = custom_health_check_config
         self.custom_runtime_config = custom_runtime_config
         self.description = description
         self.environment_variables = environment_variables
@@ -3151,6 +3502,8 @@ class CreateFunctionResponseBody(TeaModel):
             self.custom_container_config.validate()
         if self.custom_dns:
             self.custom_dns.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.custom_runtime_config:
             self.custom_runtime_config.validate()
         if self.instance_lifecycle_config:
@@ -3174,6 +3527,8 @@ class CreateFunctionResponseBody(TeaModel):
             result['customContainerConfig'] = self.custom_container_config.to_map()
         if self.custom_dns is not None:
             result['customDNS'] = self.custom_dns.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.custom_runtime_config is not None:
             result['customRuntimeConfig'] = self.custom_runtime_config.to_map()
         if self.description is not None:
@@ -3226,6 +3581,9 @@ class CreateFunctionResponseBody(TeaModel):
         if m.get('customDNS') is not None:
             temp_model = CustomDNS()
             self.custom_dns = temp_model.from_map(m['customDNS'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('customRuntimeConfig') is not None:
             temp_model = CustomRuntimeConfig()
             self.custom_runtime_config = temp_model.from_map(m['customRuntimeConfig'])
@@ -3574,6 +3932,7 @@ class CreateServiceRequest(TeaModel):
         internet_access: bool = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         service_name: str = None,
         tracing_config: TracingConfig = None,
@@ -3583,6 +3942,7 @@ class CreateServiceRequest(TeaModel):
         self.internet_access = internet_access
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.service_name = service_name
         self.tracing_config = tracing_config
@@ -3593,6 +3953,8 @@ class CreateServiceRequest(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -3612,6 +3974,8 @@ class CreateServiceRequest(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.service_name is not None:
@@ -3634,6 +3998,9 @@ class CreateServiceRequest(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('serviceName') is not None:
@@ -3656,6 +4023,7 @@ class CreateServiceResponseBody(TeaModel):
         last_modified_time: str = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         service_id: str = None,
         service_name: str = None,
@@ -3668,6 +4036,7 @@ class CreateServiceResponseBody(TeaModel):
         self.last_modified_time = last_modified_time
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.service_id = service_id
         self.service_name = service_name
@@ -3679,6 +4048,8 @@ class CreateServiceResponseBody(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -3702,6 +4073,8 @@ class CreateServiceResponseBody(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.service_id is not None:
@@ -3730,6 +4103,9 @@ class CreateServiceResponseBody(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('serviceId') is not None:
@@ -5661,6 +6037,7 @@ class GetFunctionResponseBody(TeaModel):
         created_time: str = None,
         custom_container_config: CustomContainerConfigInfo = None,
         custom_dns: CustomDNS = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         custom_runtime_config: CustomRuntimeConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
@@ -5685,6 +6062,7 @@ class GetFunctionResponseBody(TeaModel):
         self.created_time = created_time
         self.custom_container_config = custom_container_config
         self.custom_dns = custom_dns
+        self.custom_health_check_config = custom_health_check_config
         self.custom_runtime_config = custom_runtime_config
         self.description = description
         self.environment_variables = environment_variables
@@ -5708,6 +6086,8 @@ class GetFunctionResponseBody(TeaModel):
             self.custom_container_config.validate()
         if self.custom_dns:
             self.custom_dns.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.custom_runtime_config:
             self.custom_runtime_config.validate()
         if self.instance_lifecycle_config:
@@ -5731,6 +6111,8 @@ class GetFunctionResponseBody(TeaModel):
             result['customContainerConfig'] = self.custom_container_config.to_map()
         if self.custom_dns is not None:
             result['customDNS'] = self.custom_dns.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.custom_runtime_config is not None:
             result['customRuntimeConfig'] = self.custom_runtime_config.to_map()
         if self.description is not None:
@@ -5783,6 +6165,9 @@ class GetFunctionResponseBody(TeaModel):
         if m.get('customDNS') is not None:
             temp_model = CustomDNS()
             self.custom_dns = temp_model.from_map(m['customDNS'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('customRuntimeConfig') is not None:
             temp_model = CustomRuntimeConfig()
             self.custom_runtime_config = temp_model.from_map(m['customRuntimeConfig'])
@@ -6872,6 +7257,7 @@ class GetServiceResponseBody(TeaModel):
         last_modified_time: str = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         service_id: str = None,
         service_name: str = None,
@@ -6884,6 +7270,7 @@ class GetServiceResponseBody(TeaModel):
         self.last_modified_time = last_modified_time
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.service_id = service_id
         self.service_name = service_name
@@ -6895,6 +7282,8 @@ class GetServiceResponseBody(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -6918,6 +7307,8 @@ class GetServiceResponseBody(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.service_id is not None:
@@ -6946,6 +7337,9 @@ class GetServiceResponseBody(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('serviceId') is not None:
@@ -8514,6 +8908,7 @@ class ListFunctionsResponseBodyFunctions(TeaModel):
         code_size: int = None,
         created_time: str = None,
         custom_container_config: CustomContainerConfig = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
         function_id: str = None,
@@ -8536,6 +8931,7 @@ class ListFunctionsResponseBodyFunctions(TeaModel):
         self.code_size = code_size
         self.created_time = created_time
         self.custom_container_config = custom_container_config
+        self.custom_health_check_config = custom_health_check_config
         self.description = description
         self.environment_variables = environment_variables
         self.function_id = function_id
@@ -8556,6 +8952,8 @@ class ListFunctionsResponseBodyFunctions(TeaModel):
     def validate(self):
         if self.custom_container_config:
             self.custom_container_config.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.instance_lifecycle_config:
             self.instance_lifecycle_config.validate()
 
@@ -8575,6 +8973,8 @@ class ListFunctionsResponseBodyFunctions(TeaModel):
             result['createdTime'] = self.created_time
         if self.custom_container_config is not None:
             result['customContainerConfig'] = self.custom_container_config.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.description is not None:
             result['description'] = self.description
         if self.environment_variables is not None:
@@ -8622,6 +9022,9 @@ class ListFunctionsResponseBodyFunctions(TeaModel):
         if m.get('customContainerConfig') is not None:
             temp_model = CustomContainerConfig()
             self.custom_container_config = temp_model.from_map(m['customContainerConfig'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('description') is not None:
             self.description = m.get('description')
         if m.get('environmentVariables') is not None:
@@ -10195,6 +10598,7 @@ class ListServicesResponseBodyServices(TeaModel):
         last_modified_time: str = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         service_id: str = None,
         service_name: str = None,
@@ -10207,6 +10611,7 @@ class ListServicesResponseBodyServices(TeaModel):
         self.last_modified_time = last_modified_time
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.service_id = service_id
         self.service_name = service_name
@@ -10218,6 +10623,8 @@ class ListServicesResponseBodyServices(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -10241,6 +10648,8 @@ class ListServicesResponseBodyServices(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.service_id is not None:
@@ -10269,6 +10678,9 @@ class ListServicesResponseBodyServices(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('serviceId') is not None:
@@ -12354,6 +12766,85 @@ class RegisterEventSourceResponse(TeaModel):
         return self
 
 
+class ReleaseGPUInstanceHeaders(TeaModel):
+    def __init__(
+        self,
+        common_headers: Dict[str, str] = None,
+        x_fc_account_id: str = None,
+        x_fc_date: str = None,
+        x_fc_trace_id: str = None,
+    ):
+        self.common_headers = common_headers
+        self.x_fc_account_id = x_fc_account_id
+        self.x_fc_date = x_fc_date
+        self.x_fc_trace_id = x_fc_trace_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.common_headers is not None:
+            result['commonHeaders'] = self.common_headers
+        if self.x_fc_account_id is not None:
+            result['X-Fc-Account-Id'] = self.x_fc_account_id
+        if self.x_fc_date is not None:
+            result['X-Fc-Date'] = self.x_fc_date
+        if self.x_fc_trace_id is not None:
+            result['X-Fc-Trace-Id'] = self.x_fc_trace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('commonHeaders') is not None:
+            self.common_headers = m.get('commonHeaders')
+        if m.get('X-Fc-Account-Id') is not None:
+            self.x_fc_account_id = m.get('X-Fc-Account-Id')
+        if m.get('X-Fc-Date') is not None:
+            self.x_fc_date = m.get('X-Fc-Date')
+        if m.get('X-Fc-Trace-Id') is not None:
+            self.x_fc_trace_id = m.get('X-Fc-Trace-Id')
+        return self
+
+
+class ReleaseGPUInstanceResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+
+    def validate(self):
+        self.validate_required(self.headers, 'headers')
+        self.validate_required(self.status_code, 'status_code')
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        return self
+
+
 class StopStatefulAsyncInvocationHeaders(TeaModel):
     def __init__(
         self,
@@ -13185,6 +13676,7 @@ class UpdateFunctionRequest(TeaModel):
         code: Code = None,
         custom_container_config: CustomContainerConfig = None,
         custom_dns: CustomDNS = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         custom_runtime_config: CustomRuntimeConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
@@ -13204,6 +13696,7 @@ class UpdateFunctionRequest(TeaModel):
         self.code = code
         self.custom_container_config = custom_container_config
         self.custom_dns = custom_dns
+        self.custom_health_check_config = custom_health_check_config
         self.custom_runtime_config = custom_runtime_config
         self.description = description
         self.environment_variables = environment_variables
@@ -13225,6 +13718,8 @@ class UpdateFunctionRequest(TeaModel):
             self.custom_container_config.validate()
         if self.custom_dns:
             self.custom_dns.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.custom_runtime_config:
             self.custom_runtime_config.validate()
         if self.instance_lifecycle_config:
@@ -13246,6 +13741,8 @@ class UpdateFunctionRequest(TeaModel):
             result['customContainerConfig'] = self.custom_container_config.to_map()
         if self.custom_dns is not None:
             result['customDNS'] = self.custom_dns.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.custom_runtime_config is not None:
             result['customRuntimeConfig'] = self.custom_runtime_config.to_map()
         if self.description is not None:
@@ -13289,6 +13786,9 @@ class UpdateFunctionRequest(TeaModel):
         if m.get('customDNS') is not None:
             temp_model = CustomDNS()
             self.custom_dns = temp_model.from_map(m['customDNS'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('customRuntimeConfig') is not None:
             temp_model = CustomRuntimeConfig()
             self.custom_runtime_config = temp_model.from_map(m['customRuntimeConfig'])
@@ -13329,6 +13829,7 @@ class UpdateFunctionResponseBody(TeaModel):
         created_time: str = None,
         custom_container_config: CustomContainerConfig = None,
         custom_dns: CustomDNS = None,
+        custom_health_check_config: CustomHealthCheckConfig = None,
         custom_runtime_config: CustomRuntimeConfig = None,
         description: str = None,
         environment_variables: Dict[str, str] = None,
@@ -13352,6 +13853,7 @@ class UpdateFunctionResponseBody(TeaModel):
         self.created_time = created_time
         self.custom_container_config = custom_container_config
         self.custom_dns = custom_dns
+        self.custom_health_check_config = custom_health_check_config
         self.custom_runtime_config = custom_runtime_config
         self.description = description
         self.environment_variables = environment_variables
@@ -13374,6 +13876,8 @@ class UpdateFunctionResponseBody(TeaModel):
             self.custom_container_config.validate()
         if self.custom_dns:
             self.custom_dns.validate()
+        if self.custom_health_check_config:
+            self.custom_health_check_config.validate()
         if self.custom_runtime_config:
             self.custom_runtime_config.validate()
         if self.instance_lifecycle_config:
@@ -13397,6 +13901,8 @@ class UpdateFunctionResponseBody(TeaModel):
             result['customContainerConfig'] = self.custom_container_config.to_map()
         if self.custom_dns is not None:
             result['customDNS'] = self.custom_dns.to_map()
+        if self.custom_health_check_config is not None:
+            result['customHealthCheckConfig'] = self.custom_health_check_config.to_map()
         if self.custom_runtime_config is not None:
             result['customRuntimeConfig'] = self.custom_runtime_config.to_map()
         if self.description is not None:
@@ -13447,6 +13953,9 @@ class UpdateFunctionResponseBody(TeaModel):
         if m.get('customDNS') is not None:
             temp_model = CustomDNS()
             self.custom_dns = temp_model.from_map(m['customDNS'])
+        if m.get('customHealthCheckConfig') is not None:
+            temp_model = CustomHealthCheckConfig()
+            self.custom_health_check_config = temp_model.from_map(m['customHealthCheckConfig'])
         if m.get('customRuntimeConfig') is not None:
             temp_model = CustomRuntimeConfig()
             self.custom_runtime_config = temp_model.from_map(m['customRuntimeConfig'])
@@ -13586,6 +14095,7 @@ class UpdateServiceRequest(TeaModel):
         internet_access: bool = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         tracing_config: TracingConfig = None,
         vpc_config: VPCConfig = None,
@@ -13594,6 +14104,7 @@ class UpdateServiceRequest(TeaModel):
         self.internet_access = internet_access
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.tracing_config = tracing_config
         self.vpc_config = vpc_config
@@ -13603,6 +14114,8 @@ class UpdateServiceRequest(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -13622,6 +14135,8 @@ class UpdateServiceRequest(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.tracing_config is not None:
@@ -13642,6 +14157,9 @@ class UpdateServiceRequest(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('tracingConfig') is not None:
@@ -13662,6 +14180,7 @@ class UpdateServiceResponseBody(TeaModel):
         last_modified_time: str = None,
         log_config: LogConfig = None,
         nas_config: NASConfig = None,
+        oss_mount_config: OSSMountConfig = None,
         role: str = None,
         service_id: str = None,
         service_name: str = None,
@@ -13674,6 +14193,7 @@ class UpdateServiceResponseBody(TeaModel):
         self.last_modified_time = last_modified_time
         self.log_config = log_config
         self.nas_config = nas_config
+        self.oss_mount_config = oss_mount_config
         self.role = role
         self.service_id = service_id
         self.service_name = service_name
@@ -13685,6 +14205,8 @@ class UpdateServiceResponseBody(TeaModel):
             self.log_config.validate()
         if self.nas_config:
             self.nas_config.validate()
+        if self.oss_mount_config:
+            self.oss_mount_config.validate()
         if self.tracing_config:
             self.tracing_config.validate()
         if self.vpc_config:
@@ -13708,6 +14230,8 @@ class UpdateServiceResponseBody(TeaModel):
             result['logConfig'] = self.log_config.to_map()
         if self.nas_config is not None:
             result['nasConfig'] = self.nas_config.to_map()
+        if self.oss_mount_config is not None:
+            result['ossMountConfig'] = self.oss_mount_config.to_map()
         if self.role is not None:
             result['role'] = self.role
         if self.service_id is not None:
@@ -13736,6 +14260,9 @@ class UpdateServiceResponseBody(TeaModel):
         if m.get('nasConfig') is not None:
             temp_model = NASConfig()
             self.nas_config = temp_model.from_map(m['nasConfig'])
+        if m.get('ossMountConfig') is not None:
+            temp_model = OSSMountConfig()
+            self.oss_mount_config = temp_model.from_map(m['ossMountConfig'])
         if m.get('role') is not None:
             self.role = m.get('role')
         if m.get('serviceId') is not None: